Moving autoscaler crd and controller to respect default project structure
